@@ -1,11 +1,12 @@
 name: example
 description: A new Flutter project.
-version: 0.1.1
+version: 0.2.0
 
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=3.0.6 <4.0.0'
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
