@@ -1,7 +1,7 @@
 name: seatsio
 description: Seatsio SDK for Flutter.
 repository: https://github.com/SongJiaqiang/seatsio-flutter
-version: 0.3.4
+version: 0.4.0
 
 environment:
   sdk: '>=3.0.6 <4.0.0'
@@ -11,14 +11,14 @@ dependencies:
   flutter:
     sdk: flutter
 
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   built_collection: ^5.1.1
-  built_value: ^8.9.1
-  webview_flutter: ^4.7.0
+  built_value: ^8.9.3
+  webview_flutter: ^4.10.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^2.4.8
-  built_value_generator: ^8.9.1
+  build_runner: ^2.4.14
+  built_value_generator: ^8.9.3
