@@ -193,8 +193,10 @@ class SeatsioLabelBuilder
   _$SeatsioLabel _build() {
     final _$result = _$v ??
         new _$SeatsioLabel._(
-            own: own,
-            parent: parent,
+            own: BuiltValueNullFieldError.checkNotNull(
+                own, r'SeatsioLabel', 'own'),
+            parent: BuiltValueNullFieldError.checkNotNull(
+                parent, r'SeatsioLabel', 'parent'),
             section: section,
             displayedLabel: displayedLabel);
     replace(_$result);
