@@ -150,10 +150,11 @@ class SeatsioTicketTypeBuilder
   _$SeatsioTicketType _build() {
     final _$result = _$v ??
         new _$SeatsioTicketType._(
-            price: BuiltValueNullFieldError.checkNotNull(
-                price, r'SeatsioTicketType', 'price'),
-            ticketType: BuiltValueNullFieldError.checkNotNull(
-                ticketType, r'SeatsioTicketType', 'ticketType'));
+          price: BuiltValueNullFieldError.checkNotNull(
+              price, r'SeatsioTicketType', 'price'),
+          ticketType: BuiltValueNullFieldError.checkNotNull(
+              ticketType, r'SeatsioTicketType', 'ticketType'),
+        );
     replace(_$result);
     return _$result;
   }
