@@ -1,11 +1,11 @@
 import 'dart:convert';
-import 'dart:developer' as developer;
 
 import 'package:flutter/material.dart';
 import 'package:seatsio/src/assets/seatsio_html.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 import '../models/seating_chart_config.dart';
 import '../util/seatsio_js_bridge.dart';
+import '../assets/seatsio_html.dart';
 
 typedef void SeatsioWebViewCreatedCallback(SeatsioWebViewController controller);
 
@@ -83,5 +83,7 @@ class SeatsioWebViewController {
     return url.toString();
   }
 
-  Future<void> evaluateJavascript(String javascriptString) => _webViewController.runJavaScript(javascriptString);
+  Future<void> evaluateJavascript(String javascriptString) {
+    return _webViewController.runJavaScript(javascriptString);
+  }
 }
